+++ app/templates/source/js/_requireconfig.js
@@ -1,0 +1,12 @@
+requirejs({
+	"baseUrl": "./js/",
+	"noGlobal": true,
+	"paths": {
+		"logging": "../../bower_components/logging.js/logging"
+	},
+	"shim": {
+	},
+	"include": [
+		'app'
+	]
+})