+++ app/templates/source/js/app.js
@@ -1,0 +1,16 @@
+define('app', ['logging'], function(logging)
+{
+	var App = function()
+	{
+		logging.applyLogging(this, 'App');
+
+		this.initializeEventListeners();
+	};
+
+	App.prototype.initializeEventListeners = function()
+	{
+
+	};
+
+	new App();
+});