+++ app/templates/source/js/modernizr/positionStickyTest.js
@@ -1,0 +1,22 @@
+/*!
+{
+  "name": "CSS position: sticky",
+  "property": "csspositionsticky",
+  "tags": ["css"],
+  "notes": [
+	"using position:sticky on anything but top aligned elements is buggy in Chrome < 37 and iOS <=7+", {
+	"name": "Chrome bug report",
+	"href":"https://code.google.com/p/chromium/issues/detail?id=322972"
+  }]
+}
+!*/
+Modernizr.addTest('csspositionsticky', function () {
+	var prop = 'position:';
+	var value = 'sticky';
+	var el = document.createElement('modernizr');
+	var mStyle = el.style;
+	
+	mStyle.cssText = prop + Modernizr._prefixes.join(value + ';' + prop).slice(0, -prop.length);
+	
+	return mStyle.position.indexOf(value) !== -1;
+});